refactor CreateBlobFileIO
